+++ WebStreamer/__init__.py
@@ -1,10 +1,12 @@
 # This file is a part of TG-
 # Coding : Jyothis Jayanth [@EverythingSuckz]
-
-
-import time
+from time import time
 from .vars import Var
-from WebStreamer.bot.clients import StreamBot
+from WebStreamer.bot import StreamBot
+
+import asyncio
+
+bot_loop = asyncio.get_event_loop()
 
 __version__ = 2.2
-StartTime = time.time()
+StartTime = time()
